Refactoring of offer details
